+++ constant/product.go
@@ -1,0 +1,8 @@
+package constant
+
+const (
+	StatusApproved   = "APPROVED"
+	StatusWaiting    = "WAITING"
+	StatusRejected   = "REJECTED"
+	StatusSubscribed = "SUBSCRIBED"
+)