+++ datastruct/brand.go
@@ -6,6 +6,7 @@ type Brand struct {
 	ID         uint64    `gorm:"column:id;primaryKey;autoIncrement" json:"id"`
 	Name       string    `gorm:"column:name" json:"name"`
 	CategoryID uint64    `gorm:"column:category_id" json:"category_id"`
+	Image      string    `gorm:"column:image" json:"image"`
 	CreatedAt  time.Time `gorm:"column:created_at" json:"created_at"`
 	UpdatedAt  time.Time `gorm:"column:updated_at" json:"updated_at"`
 }